Resolve circular dependencies
- Add a Webpack plugin to our dev config that checks for circular dependencies
- Resolve circular dependencies involving resetEventsPolling (which was used inside the store
and itself relied on the store)
- Resolve Lish/Glish dependency chain (Glish was importing methods from Lish, which in turn
imported Glish)
- Resolve reportException chain (moved handleUnauthedErrors to its own utility)
