+++ packages/manager/src/events.ts
@@ -14,16 +14,13 @@ import { Event } from 'linode-js-sdk/lib/account';
 import { Subject } from 'rxjs/Subject';
 import { DISABLE_EVENT_THROTTLE } from 'src/constants';
 import store from 'src/store';
+import {
+  setPollingInterval,
+  setRequestDeadline
+} from 'src/store/events/event.actions';
 import { getEvents } from 'src/store/events/event.request';
 
 export const events$ = new Subject<Event>();
-
-export let eventRequestDeadline = Date.now();
-
-/**
- * The current iteration of the poll. 1, 2, 4, 8, 16.
- */
-export let pollIteration = 1;
 
 /**
  * The lowest interval at which to make a request. This is later multiplied by the pollIteration
@@ -34,8 +31,8 @@ export const INTERVAL: number = 1000;
 let inProgress = false;
 
 export const resetEventsPolling = (newPollIteration = 1) => {
-  eventRequestDeadline = Date.now() + INTERVAL * newPollIteration;
-  pollIteration = newPollIteration;
+  store.dispatch(setRequestDeadline(Date.now() + INTERVAL * newPollIteration));
+  store.dispatch(setPollingInterval(newPollIteration));
 };
 
 export const requestEvents = () => {
@@ -57,7 +54,10 @@ export const requestEvents = () => {
 export const startEventsInterval = () =>
   setInterval(
     () => {
+      const state = store.getState();
       const now = Date.now();
+      const pollIteration = state.events.pollingInterval || 1;
+      const eventRequestDeadline = state.events.requestDeadline;
       if (now > eventRequestDeadline) {
         /**
          * If we're waiting on a request, set reset the pollIteration and return to prevent
@@ -75,13 +75,13 @@ export const startEventsInterval = () =>
           /*
            * If throttling is disabled manually set the timeout so tests wait to query the mock data store.
            */
-          eventRequestDeadline = now + 500;
+          store.dispatch(setRequestDeadline(now + 500));
         } else {
           const timeout = INTERVAL * pollIteration;
           /** Update the dealing */
-          eventRequestDeadline = now + timeout;
+          store.dispatch(setRequestDeadline(now + timeout));
           /* Update the iteration to a maximum of 16. */
-          pollIteration = Math.min(pollIteration * 2, 16);
+          store.dispatch(setPollingInterval(Math.min(pollIteration * 2, 16)));
         }
       }
     },
