+++ packages/manager/src/store/middleware/combineEventsMiddleware.ts
@@ -1,13 +1,12 @@
 import { compose, equals, uniqWith } from 'ramda';
 import { Middleware } from 'redux';
-import { resetEventsPolling } from 'src/events';
 import {
   isEntityEvent,
   isInProgressEvent
 } from 'src/store/events/event.helpers';
 import { EventHandler } from 'src/store/types';
 import { isType } from 'typescript-fsa';
-import { addEvents } from '../events/event.actions';
+import { addEvents, setPollingInterval } from '../events/event.actions';
 import { ExtendedEvent } from '../events/event.helpers';
 
 const eventsMiddlewareFactory = (
@@ -54,7 +53,7 @@ const eventsMiddlewareFactory = (
        */
       if (isInProgressEvent(event)) {
         // If the event is in_progress, we poll more aggressively
-        resetEventsPolling();
+        setPollingInterval(1);
       }
     }
   }
